requirements.txt
fastapi~=0.115.6
black==24.10.0
flake8==7.1.1
pre_commit==4.1.0
mypy==1.14.1
pytest==8.3.4
uvicorn==0.34.0
pytest-cov==6.0.0
requests==2.32.3
pydantic~=2.10.5
python-dotenv~=1.0.1
